Created: curvefit_automated_random_pguesses
I have automated the guess parameters by creating a function that takes the true parameters, picks a random number within a threshold of the true value that you define, and rounds it to an interval that you pick.
The code then creates noise for the data and then does curve fit many times with different guesses.